Always follow rules-of-hooks
Whenever the linter warns about compliance to rule-of-hooks, it generally means there is a potential bug in the code.
Must of the time, it's a known case in which the ref is constant and will never change (see history) BUT it looks better to always list all of those deps to avoid nasty bugs.